requirements.txt
django
channels
channels-redis
django-imagekit
Pillow
requests